test(signals): relocate provider unit tests into the package
