+++ packages/signals/tests/providers/coverage-trend.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
-import { coverageTrendProvider } from '../../../../src/server/signals/providers/coverage-trend';
-import { SignalTimelineStore } from '../../../../src/server/signals/timeline-store';
-import type { SignalContext, CommandRunner } from '../../../../src/server/signals/types';
+import { coverageTrendProvider } from '../../src/providers/coverage-trend';
+import { SignalTimelineStore } from '../../src/timeline-store';
+import type { SignalContext, CommandRunner } from '../../src/types';
 
 const US = '\x1f'; // unit (field) separator within a record
